feat(teams): add update functionality for teams (#238)
Introduced `updateTeam` mutation and CLI command to update team
details such as email. Updated schema and GraphQL queries to support
new fields.

+++ cli/src/command/auth/transfer.rs
@@ -11,7 +11,7 @@ pub struct TransferArgs {
     pub team: String,
 
     #[arg(help = "The amount to transfer.", value_name = "amount")]
-    pub amount: f64,
+    pub amount: i64,
 }
 
 impl TransferArgs {
